pyproject.toml
[project]
name = "ucar-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=6.0.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.6",
    "httpx>=0.28.1",
    "langchain>=1.2.9",
    "langchain-openai>=1.1.8",
    "openai>=2.20.0",
    "pandas>=2.0.0,<3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.41.0",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.40.0",
    "uvicorn[standard]>=0.40.0",
    "webdriver-manager>=4.0.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]
    